camel-kamelet-main - Cleanup code
